Rewording (#439)
The problem was that for zwe-mat* the application would crash when loading a the record component.
The cause was and reason why it affected only zwe-mat* is that it had one level in the geodata, so the  `singularise` function got called with a undefined value making it to crash. The simple solution is to use a different wording and avoid the  `singularise` function for now.